More type hinting fix-ups.

+++ arcade/arcade_types.py
@@ -5,5 +5,5 @@ from typing import Union
 RGB = Union[Tuple[int, int, int], List[int]]
 RGBA = Union[Tuple[int, int, int, int], List[int]]
 Color = Union[RGB, RGBA]
-Point = Union[Tuple[int, int], List[int]]
+Point = Union[Tuple[float, float], List[float]]
 PointList = Union[Tuple[Point, ...], List[Point]]
